# colexec: use execgen directives for and_or op