… pkg/sql/colexec/and_or_projection.eg.go