This is a pretty good example of how we can use execgen:inline/template
even on functions that need type templating, which I didn't realize
before. It works fine because the template variables are left alone.

This allows some nice flexibility. For helper functions that needed
templating but didn't want inlining (there are many like this that are
not in inner loops, throughout the codebase), we needed to use the full
text/template inlining hammer which made enormous function bodies for no
good reason.

Now, we can use execgen:template on its own when inlining isn't
necessary, and add inlining only for the function bodies that need to
live in inner loops.

The performance differences here I'm pretty sure are just random noise.
I can't justify why we'd get 2x speed increase in some cases at all.

```
name                                              old time/op    new time/op    delta
LogicalProjOp/AND,useSel=true,hasNulls=true-24      9.62µs ± 1%   11.52µs ± 3%   +19.72%  (p=0.000 n=10+10)
LogicalProjOp/AND,useSel=true,hasNulls=false-24     4.45µs ± 3%    4.97µs ± 1%   +11.78%  (p=0.000 n=9+9)
LogicalProjOp/AND,useSel=false,hasNulls=true-24     9.52µs ± 1%   10.59µs ± 1%   +11.15%  (p=0.000 n=10+10)
LogicalProjOp/AND,useSel=false,hasNulls=false-24    10.5µs ± 4%     4.9µs ± 2%   -53.48%  (p=0.000 n=10+10)
LogicalProjOp/OR,useSel=true,hasNulls=true-24       12.0µs ± 4%    11.0µs ± 3%    -8.38%  (p=0.000 n=10+10)
LogicalProjOp/OR,useSel=true,hasNulls=false-24      3.78µs ± 2%    4.36µs ± 2%   +15.14%  (p=0.000 n=10+9)
LogicalProjOp/OR,useSel=false,hasNulls=true-24      10.1µs ± 2%    11.2µs ± 2%   +11.26%  (p=0.000 n=10+10)
LogicalProjOp/OR,useSel=false,hasNulls=false-24     6.00µs ± 2%    4.37µs ± 4%   -27.23%  (p=0.000 n=10+9)

name                                              old speed      new speed      delta
LogicalProjOp/AND,useSel=true,hasNulls=true-24     851MB/s ± 1%   711MB/s ± 3%   -16.44%  (p=0.000 n=10+10)
LogicalProjOp/AND,useSel=true,hasNulls=false-24   1.84GB/s ± 3%  1.65GB/s ± 1%   -10.56%  (p=0.000 n=9+9)
LogicalProjOp/AND,useSel=false,hasNulls=true-24    860MB/s ± 1%   773MB/s ± 1%   -10.16%  (p=0.000 n=10+9)
LogicalProjOp/AND,useSel=false,hasNulls=false-24   783MB/s ± 4%  1683MB/s ± 2%  +114.92%  (p=0.000 n=10+10)
LogicalProjOp/OR,useSel=true,hasNulls=true-24      682MB/s ± 4%   744MB/s ± 3%    +9.12%  (p=0.000 n=10+10)
LogicalProjOp/OR,useSel=true,hasNulls=false-24    2.16GB/s ± 2%  1.88GB/s ± 2%   -13.13%  (p=0.000 n=10+9)
LogicalProjOp/OR,useSel=false,hasNulls=true-24     815MB/s ± 2%   732MB/s ± 2%   -10.12%  (p=0.000 n=10+10)
LogicalProjOp/OR,useSel=false,hasNulls=false-24   1.36GB/s ± 1%  1.86GB/s ± 5%   +36.64%  (p=0.000 n=10+10)
```

Release note: None