@jordanlewis 2020-07-09
I decided it's not worth it to refactor into structs, because the text will still be as long - you'll still have to put everything into the struct and then all the accesses will be longer. I did your suggestion with line length.