@asubiotto 2020-07-09
Are you sure the benchmarks are just random noise? It seems like a pretty huge difference and would indicate that either there is a real performance regression, or the benchmark is broken. Might be worth investigating a bit.